fix: move HTTP client constructor to client.go and fix lint errors
